Bluetooth : Remove out of range devices after ScanComplete
Change to remove device from UI when Bluetooth of remote
device is off or not discoverable anymore. Without this
change a device that discovered previously is available
in available devices list always, though the remote device
is not discoverable anymore. The current change updates
available devices list on scan complete removing the
devices that are available in cached devices and are not
disvorable anymore.

Change-Id: I97809fcafd904271121e26741e50a9b61f0e0664

Conflicts:
	src/com/android/settings/bluetooth/CachedBluetoothDevice.java
	src/com/android/settings/bluetooth/CachedBluetoothDeviceManager.java
	src/com/android/settings/bluetooth/DeviceListPreferenceFragment.java